refactor: rework utils structure, update imports
