+++ cards-ad/src/components/game/game-state/game-state.ts
@@ -1,4 +1,6 @@
-import { GameState, gameTime } from "src/utils";
+import { gameTime } from "src/utils/constants";
+import { GameState } from "src/utils/types";
+
 
 
 let state: GameState = {
