+++ cards-ad/src/components/game/init-game/init-game.ts
@@ -1,9 +1,11 @@
 
 import { dealCards } from "src/components/deck/deck";
-import { gameTime, Messages } from "src/utils";
 import { updateGameState } from "../game-state/game-state";
 import { startTimer } from "../timer/timer";
 import { showMessage } from "../show-message/show-message";
+
+import { gameTime } from "src/utils/constants";
+import { Messages } from "src/utils/types";
 
 
 export async function initGame(): Promise<void> {
