+++ cards-ad/src/components/layout/messages/messages.ts
@@ -1,5 +1,6 @@
 import './messages.css';
 
+
 export function createMessageWindow(root: HTMLElement): void {
   const messageElement = document.createElement('p');
   messageElement.classList.add('message-window');
