+++ cards-ad/src/utils/helpers.ts
@@ -1,0 +1,4 @@
+
+export function debounce(ms: number): Promise<void> {
+  return new Promise(resolve => setTimeout(resolve, ms));
+}
